requirements.txt
mysql-connector==2.2.9
flask==1.1.1
